+++ packages/post-processors/src/index.ts
@@ -3,3 +3,4 @@ export * from './constants/index.js';
 export * from './enums/index.js';
 export * from './interfaces/index.js';
 export * from './types/index.js';
+export * from './common/index.js';