+++ packages/workers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/outbox@0.8.0
+
 ## 1.2.3
 
 ### Patch Changes